requirements.txt
CairoSVG==2.6.0
PyYAML==6.0
svgwrite==1.4.3
